docs/requirements.txt
guzzle_sphinx_theme
pygments-mathematica
sphinxcontrib-bibtex
pybtex